cumulative updates

+++ article.php
@@ -1,4 +1,7 @@
 <?php
+
+session_start();
+
 // PAGE VARIABLES
 $pageTitle = 'DJ Bostock - Blog';
 
@@ -41,8 +44,10 @@ $article = getArticle($id, $conn);
                 <?php else: ?>
                     <h2><?= htmlspecialchars($article['title']); ?></h2>
                     <p><?= htmlspecialchars($article['content']); ?></p>
-                    <p><a href="edit-article.php?id=<?= $article['id']; ?>">Edit Article</a></p>
-                    <p><a href="delete-article.php?id=<?= $article['id']; ?>">Delete Article</a></p>
+                    <?php if ($_SESSION['is_logged_in']): ?>
+                        <p><a href="edit-article.php?id=<?= $article['id']; ?>">Edit Article</a></p>
+                        <p><a href="delete-article.php?id=<?= $article['id']; ?>">Delete Article</a></p>
+                    <?php endif; ?>
                 <?php endif; ?>
             </article>
